requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.6
colorama==0.4.6
ffmpeg-python==0.2.0
filelock==3.25.2
fsspec==2026.2.0
future==1.0.0
idna==3.11
Jinja2==3.1.6
llvmlite==0.46.0
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.6.1
numba==0.64.0
numpy==2.4.3
openai-whisper==20250625
RapidFuzz==3.14.3
regex==2026.2.28
requests==2.32.5
sympy==1.14.0
tiktoken==0.12.0
torch==2.10.0
tqdm==4.67.3
typing_extensions==4.15.0
urllib3==2.6.3
webrtcvad-wheels==2.0.14
